Changed to lowercase package names in AbyssalCreatures
