+++ AbyssalCreatures/src/main/java/com/offz/spigot/abyssalcreatures/mobs/flying/Benikuchinawa.java
@@ -1,4 +1,4 @@
-package com.offz.spigot.abyssalcreatures.Mobs.Flying;
+package com.offz.spigot.abyssalcreatures.mobs.flying;
 
 import com.offz.spigot.mobzy.mobs.behaviours.HitBehaviour;
 import com.offz.spigot.mobzy.mobs.types.FlyingMob;
