+++ AbyssalCreatures/src/main/java/com/offz/spigot/abyssalcreatures/mobs/flying/Cyatoria.java
@@ -1,13 +1,13 @@
-package com.offz.spigot.abyssalcreatures.Mobs.Flying;
+package com.offz.spigot.abyssalcreatures.mobs.flying;
 
 import com.offz.spigot.mobzy.mobs.behaviours.HitBehaviour;
 import com.offz.spigot.mobzy.mobs.types.FlyingMob;
 import com.offz.spigot.mobzy.pathfinders.flying.PathfinderGoalDiveOnTargetAttack;
 import net.minecraft.server.v1_13_R2.World;
 
-public class Dosetori extends FlyingMob implements HitBehaviour {
-    public Dosetori(World world) {
-        super(world, "Dosetori");
+public class Cyatoria extends FlyingMob implements HitBehaviour {
+    public Cyatoria(World world) {
+        super(world, "Cyatoria");
         setSize(3F, 2F);
     }
