+++ AbyssalCreatures/src/main/java/com/offz/spigot/abyssalcreatures/mobs/flying/Dosetori.java
@@ -1,14 +1,14 @@
-package com.offz.spigot.abyssalcreatures.Mobs.Flying;
+package com.offz.spigot.abyssalcreatures.mobs.flying;
 
 import com.offz.spigot.mobzy.mobs.behaviours.HitBehaviour;
 import com.offz.spigot.mobzy.mobs.types.FlyingMob;
 import com.offz.spigot.mobzy.pathfinders.flying.PathfinderGoalDiveOnTargetAttack;
 import net.minecraft.server.v1_13_R2.World;
 
-public class Madokajack extends FlyingMob implements HitBehaviour {
-    public Madokajack(World world) {
-        super(world, "Madokajack");
-        setSize(4.5f, 3f);
+public class Dosetori extends FlyingMob implements HitBehaviour {
+    public Dosetori(World world) {
+        super(world, "Dosetori");
+        setSize(3F, 2F);
     }
 
     @Override
