+++ AbyssalCreatures/src/main/java/com/offz/spigot/abyssalcreatures/mobs/hostile/Ottobas.java
@@ -1,14 +1,15 @@
-package com.offz.spigot.abyssalcreatures.Mobs.Hostile;
+package com.offz.spigot.abyssalcreatures.mobs.hostile;
 
 import com.offz.spigot.mobzy.mobs.behaviours.HitBehaviour;
 import com.offz.spigot.mobzy.mobs.types.HostileMob;
 import com.offz.spigot.mobzy.pathfinders.PathfinderGoalWalkingAnimation;
 import net.minecraft.server.v1_13_R2.World;
 
-public class Tesuchi extends HostileMob implements HitBehaviour {
-    public Tesuchi(World world) {
-        super(world, "Tesuchi");
-        this.setSize(0.6F, 0.6F);
+public class Ottobas extends HostileMob implements HitBehaviour {
+    public Ottobas(World world) {
+        super(world, "Ottobas");
+        this.setSize(2F, 3F);
+
     }
 
     @Override
