+++ AbyssalCreatures/src/main/java/com/offz/spigot/abyssalcreatures/mobs/passive/Ashimite.java
@@ -1,17 +1,16 @@
-package com.offz.spigot.abyssalcreatures.Mobs.Passive;
+package com.offz.spigot.abyssalcreatures.mobs.passive;
 
 import com.offz.spigot.mobzy.mobs.behaviours.HitBehaviour;
 import com.offz.spigot.mobzy.mobs.types.PassiveMob;
 import com.offz.spigot.mobzy.pathfinders.PathfinderGoalWalkingAnimation;
 import net.minecraft.server.v1_13_R2.EntityAgeable;
 import net.minecraft.server.v1_13_R2.EntityHuman;
-import net.minecraft.server.v1_13_R2.PathfinderGoalAvoidTarget;
 import net.minecraft.server.v1_13_R2.World;
 
-public class Makihige extends PassiveMob implements HitBehaviour {
-    public Makihige(World world) {
-        super(world, "Makihige");
-        setSize(2, 2);
+public class Ashimite extends PassiveMob implements HitBehaviour {
+    public Ashimite(World world) {
+        super(world, "Ashimite");
+        setSize(2f, 2);
     }
 
     @Override
@@ -23,10 +22,9 @@ public class Makihige extends PassiveMob implements HitBehaviour {
     public void createPathfinders() {
         super.createPathfinders();
         goalSelector.a(0, new PathfinderGoalWalkingAnimation(this, getStaticBuilder().getModelID()));
-        goalSelector.a(1, new PathfinderGoalAvoidTarget<>(this, EntityHuman.class, 8.0F, 1D, 1D));
     }
 
     public PassiveMob createChild(EntityAgeable entityageable) {
-        return new Makihige(this.world);
+        return new Ashimite(this.world);
     }
 }
